==> data_preprocess/1-ProcessUserItem.ipynb <==
!which python

import spacy
import nltk
import re
import json
import pandas as pd
import os
import numpy as np

# Read Data

# Load original dataset
dir_path = '../Dataset/ratebeer/original_dataset'
# Load train dataset
train_review = []
cnt = 0
file_path = os.path.join(dir_path, 'train_no_duplicate.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        train_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 100000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading train dataset, totally {} lines.'.format(len(train_review)))
# Load test dataset
test_review = []
cnt = 0
file_path = os.path.join(dir_path, 'test_no_duplicate.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        test_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 10000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading test dataset, totally {} lines.'.format(len(test_review)))

## Convert list data to pandas dataframe

df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
df_test_data = pd.DataFrame(test_review, columns=['item', 'user', 'rating', 'review'])

df_train_data

train_review[0][3]

for idx, row in df_train_data.iterrows():
    print(row['item'])
    print(type(row['item']))
    break

## Aggregate all the relevant sentences for each user/item in the train dataset

## Filter User/Item

"""
The requirement of being a satisfactory dataframe is that for each user / item in the dataset, 
there should be at least 20 reviews and no more than 500 reviews.
"""
lower_thres = 30
upper_thres = 300
def is_satisfy_df(df_review):
    # group by user
    group_by_user = df_review.groupby('user')
    user_selected = set()
    user_review_morethanupper = set()
    user_review_lessthanlower = set()
    for user_df_chunk in list(group_by_user):
        user_id = int(user_df_chunk[0])
        user_df = user_df_chunk[1]
        if len(user_df) > upper_thres:
            user_review_morethanupper.add(user_id)
        elif len(user_df) < lower_thres:
            user_review_lessthanlower.add(user_id)
        else:
            user_selected.add(user_id)
    print("user_selected: {0} \t user(>upper): {1} \t user(<lower): {2}".format(len(user_selected), len(user_review_morethanupper), len(user_review_lessthanlower)))
    # group by item
    group_by_item = df_review.groupby('item')
    item_selected = set()
    item_review_morethanupper = set()
    item_review_lessthanlower = set()
    for item_df_chunk in list(group_by_item):
        item_id = int(item_df_chunk[0])
        item_df = item_df_chunk[1]
        if len(item_df) > upper_thres:
            item_review_morethanupper.add(item_id)
        elif len(item_df) < lower_thres:
            item_review_lessthanlower.add(item_id)
        else:
            item_selected.add(item_id)
    print("item_selected: {0} \t item(>upper): {1} \t item(<lower): {2}".format(len(item_selected), len(item_review_morethanupper), len(item_review_lessthanlower)))
    # whether this dataframe has rare/popular user/item
    if len(user_review_morethanupper) == 0 and len(user_review_lessthanlower) == 0:
        if len(item_review_morethanupper) == 0 and len(item_review_lessthanlower) == 0:
            return True
        else:
            return False
    else:
        return False

iter_cnt = 0
while True:
    # group by user
    group_by_user = df_train_data.groupby('user')
    # Loop through all the dataframe for each user
    user_selected = set()
    user_review_morethan500 = set()
    user_review_lessthan20 = set()
    user_num_review = []
    for user_df_chunk in list(group_by_user):
        user_id = int(user_df_chunk[0])
        user_df = user_df_chunk[1]
        user_num_review.append(len(user_df))
        if len(user_df) > upper_thres:
            user_review_morethan500.add(user_id)
        elif len(user_df) < lower_thres:
            user_review_lessthan20.add(user_id)
        else:
            user_selected.add(user_id)
    print("Start Drop")
    # Filter user
    # for idx, row in df_train_data.iterrows():
    #     if int(row['user']) not in user_selected:
    #         df_train_data.drop(idx, inplace=True)
    cur_train_review = []
    for row_data in train_review:
        # if the user_id is in the selected user set or not
        if int(row_data[1]) in user_selected:
            cur_train_review.append(row_data)
    train_review = cur_train_review
    # convert list to dataframe
    df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
    # check whether this dataframe satisfies the requirement
    if is_satisfy_df(df_train_data):
        break
    print('Finish User')
    # else, keep going
    # group by item
    group_by_item = df_train_data.groupby('item')
    # Loop through all the dataframe for each item
    item_selected = set()
    item_review_morethan500 = set()
    item_review_lessthan20 = set()
    item_num_review = []
    for item_df_chunk in list(group_by_item):
        item_id = int(item_df_chunk[0])
        item_df = item_df_chunk[1]
        item_num_review.append(len(item_df))
        if len(item_df) > upper_thres:
            item_review_morethan500.add(item_id)
        elif len(item_df) < lower_thres:
            item_review_lessthan20.add(item_id)
        else:
            item_selected.add(item_id)
    print("Start Drop")
    # Filter item
    # for idx, row in df_train_data.iterrows():
    #     if int(row['item']) not in item_selected:
    #         df_train_data.drop(idx, inplace=True)
    cur_train_review = []
    for row_data in train_review:
        # if the item_id is in the selected user set or not
        if int(row_data[0]) in item_selected:
            cur_train_review.append(row_data)
    train_review = cur_train_review
    # convert list to dataframe
    df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
    # check whether this dataframe satisfies the requirement
    if is_satisfy_df(df_train_data):
        break
    print('Finish Item')
    iter_cnt += 1
    if iter_cnt % 10 == 0:
        print("{} iterations of filter".format(iter_cnt))


df_train_data

### Check Whether the Filtered Dataframe Satisfies the Requirement

# Check User
# group by user
group_by_user = df_train_data.groupby('user')
user_selected = set()
user_review_morethan500 = set()
user_review_lessthan20 = set()
user_num_reviews = []
for user_df_chunk in list(group_by_user):
    user_id = int(user_df_chunk[0])
    user_df = user_df_chunk[1]
    if len(user_df) > upper_thres:
        user_review_morethan500.add(user_id)
    elif len(user_df) < lower_thres:
        user_review_lessthan20.add(user_id)
    else:
        user_selected.add(user_id)
    user_num_reviews.append(len(user_df))


print("Number of user selected: {}".format(len(user_selected)))
print("Number of user with more than upper reviews: {}".format(len(user_review_morethan500)))
print("Number of user with less than lower reviews: {}".format(len(user_review_lessthan20)))

print("Mean number of reviews per user: {}".format(np.mean(user_num_reviews)))
print("Max number of reviews per user: {}".format(np.max(user_num_reviews)))
print("Min number of reviews per user: {}".format(np.min(user_num_reviews)))

# Check Item
# group by item
group_by_item = df_train_data.groupby('item')
item_selected = set()
item_review_morethan500 = set()
item_review_lessthan20 = set()
item_num_reviews = []
for item_df_chunk in list(group_by_item):
    item_id = int(item_df_chunk[0])
    item_df = item_df_chunk[1]
    if len(item_df) > upper_thres:
        item_review_morethan500.add(item_id)
    elif len(item_df) < lower_thres:
        item_review_lessthan20.add(item_id)
    else:
        item_selected.add(item_id)
    item_num_reviews.append(len(item_df))


print("Number of item selected: {}".format(len(item_selected)))
print("Number of item with more than upper reviews: {}".format(len(item_review_morethan500)))
print("Number of item with less than lower reviews: {}".format(len(item_review_lessthan20)))

print("Mean number of reviews per item: {}".format(np.mean(item_num_reviews)))
print("Max number of reviews per item: {}".format(np.max(item_num_reviews)))
print("Min number of reviews per item: {}".format(np.min(item_num_reviews)))

is_satisfy_df(df_train_data)

for idx, row in df_train_data.iterrows():
    print(row)
    print(row['item'])
    print(type(row['item']))
    print(type(row['user']))
    print(type(row['rating']))
    print(type(row['review']))
    break

## Save the Filtered Train Set into Json File

df_train_data

# Save the dataframe to json file
with open('../Dataset/ratebeer/medium/train_review_filtered.json', 'w') as f1:
    for idx, row in df_train_data.iterrows():
        row_dict = {'user': row['user'], 'item': row['item'], 'rating': row['rating'], 'review': row['review']}
        # dump this dict into file
        json.dump(row_dict, f1)
        f1.write("\n")

## Filter Test set

df_test_data

len(test_review)

test_review_filter = []
for test_data in test_review:
    item_id = int(test_data[0])
    user_id = int(test_data[1])
    if user_id in user_selected and item_id in item_selected:
        # add this into the filtered test set
        test_review_filter.append(test_data)

len(test_review_filter)

df_test_data_filter = pd.DataFrame(test_review_filter, columns=['item', 'user', 'rating', 'review'])

df_test_data_filter

### Save Filtered Test Set into Json File

# Save the dataframe to json file
with open('../Dataset/ratebeer/medium/test_review_filtered.json', 'w') as f1:
    for idx, row in df_test_data_filter.iterrows():
        row_dict = {'user': row['user'], 'item': row['item'], 'rating': row['rating'], 'review': row['review']}
        # dump this dict into file
        json.dump(row_dict, f1)
        f1.write("\n")
