==> data_preprocess/2-feature2id.ipynb <==
!which python

import spacy
import nltk
import re
import json
import pandas as pd
import os
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer,TfidfTransformer

# Read Data

dir_path = '../Dataset/ratebeer/large_500'
# Load train dataset
train_review = []
cnt = 0
file_path = os.path.join(dir_path, 'train_review_filtered.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        train_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 10000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading train dataset, totally {} lines.'.format(len(train_review)))
# Load test dataset
test_review = []
cnt = 0
file_path = os.path.join(dir_path, 'test_review_filtered.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        test_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 10000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading test dataset, totally {} lines.'.format(len(test_review)))

## Convert List Data to Pandas Dataframe

df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
df_test_data = pd.DataFrame(test_review, columns=['item', 'user', 'rating', 'review'])

df_train_data

print("number of user in trainset: {}".format(len(list(df_train_data['user'].unique()))))
print("number of item in trainset: {}".format(len(list(df_train_data['item'].unique()))))

def catDoc(textlist):
    res = []
    for tlist in textlist:
        res.extend(tlist)
    return res

def calTFidf(text):
    # FIXED: Omit STOP Words
    vectorizer = CountVectorizer(lowercase=True, stop_words='english')
    wordcount = vectorizer.fit_transform(text)

    print("wordcount: ", wordcount.shape)
    tf_idf_transformer = TfidfTransformer()
    tfidf_matrix = tf_idf_transformer.fit_transform(wordcount)
    return vectorizer, tfidf_matrix

# Load Reviews

documents = []
for idx, row in df_train_data.iterrows():
    text = row['review']
    documents.append(text)

len(documents)

# Compute TF-IDF Matrix

vectorizer, tfidf_matrix = calTFidf(documents)

print("The number of example is {0}, and the TFIDF vocabulary size is {1}".format(
    len(documents), len(vectorizer.vocabulary_)))

## Compute Mean TF-IDF score for each word

word_tfidf = np.array(tfidf_matrix.mean(0))

word_tfidf.shape

word_tfidf[0][:20]

## Sort Words According to Their Mean TF-IDF value

word_order = np.argsort(-word_tfidf[0])

word_order.shape

word_order[:20]

import string
punct = string.punctuation

vocab2id = {}
id2vocab = {}

id2word = vectorizer.get_feature_names()
feature_word_num = 2000
feature_word_idx = 0

cnt = 0
for idx in word_order:
    # get the word
    word = id2word[idx]
    # check if this word is a number/punctuation
    if word.isdigit() or (word in punct):
        print(word)
    else:
        # write word into vocab2id mapping
        vocab2id[word] = str(len(vocab2id))
        # write word into id2vocab mapping
        id2vocab[str(cnt)] = word
        cnt += 1
        if cnt == feature_word_num:
            break

## Save the feature vocabulary to json file

len(vocab2id)

len(id2vocab)

with open('../Dataset/ratebeer/large_500/train/feature/feature2id.json', 'w') as f:
    json.dump(vocab2id, f)

with open('../Dataset/ratebeer/large_500/train/feature/id2feature.json', 'w') as f:
    json.dump(id2vocab, f)

key_cnt = 0
stack_key = []
for key, value in vocab2id.items():
    if key_cnt == 10:
        # write this into file
        with open('../Dataset/ratebeer/large_500/train/feature/features.txt', 'a') as f:
            key_text = '\t'.join(stack_key)
            f.write(key_text)
            f.write('\n')
        key_cnt = 0
        stack_key = []
    stack_key.append(key)
    key_cnt += 1

if len(stack_key) != 0:
    with open('../Dataset/ratebeer/large_500/train/feature/features.txt', 'a') as f:
        key_text = '\t'.join(stack_key)
        f.write(key_text)
        f.write('\n')
