==> data_preprocess/8-ConstructTestDatawithProxySentence.ipynb <==
!which python

import json
import pandas as pd
import os
import numpy as np

dir_path = '../Dataset/ratebeer/small_500'
# Load train dataset
train_review = []
cnt = 0
file_path = os.path.join(dir_path, 'train_review_filtered.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        train_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 100000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading train dataset, totally {} lines.'.format(len(train_review)))
# Load test dataset
test_review = []
cnt = 0
file_path = os.path.join(dir_path, 'test_review_filtered.json')
with open(file_path) as f:
    for line in f:
        line_data = json.loads(line)
        user_id = line_data['user']
        item_id = line_data['item']
        rating = line_data['rating']
        review = line_data['review']
        test_review.append([item_id, user_id, rating, review])
        cnt += 1
        if cnt % 10000 == 0:
            print('{} lines loaded.'.format(cnt))
print('Finish loading test dataset, totally {} lines.'.format(len(test_review)))

df_train_data = pd.DataFrame(train_review, columns=['item', 'user', 'rating', 'review'])
df_test_data = pd.DataFrame(test_review, columns=['item', 'user', 'rating', 'review'])

train_user_id_set = set(df_train_data['user'].unique())
train_item_id_set = set(df_train_data['item'].unique())

print("number of user on train set: {}".format(len(train_user_id_set)))
print("number of item on train set: {}".format(len(train_item_id_set)))

# Sentence2id and id2sentence
with open('../Dataset/ratebeer/small_500/train/sentence/sentence2id.json', 'r') as f:
    sent_to_id = json.load(f)
with open('../Dataset/ratebeer/small_500/train/sentence/id2sentence.json', 'r') as f:
    id_to_sent = json.load(f)

len(id_to_sent)

# Sentence2id and id2sentence
with open('../Dataset/ratebeer/small_500/test/sentence/sentence2id.json', 'r') as f:
    testset_sent_to_id = json.load(f)
with open('../Dataset/ratebeer/small_500/test/sentence/id2sentence.json', 'r') as f:
    testset_id_to_sent = json.load(f)

# validset Sentence2id and id2sentence
with open('../Dataset/ratebeer/small_500/valid/sentence/sentence2id.json', 'r') as f:
    validset_sent_to_id = json.load(f)
with open('../Dataset/ratebeer/small_500/valid/sentence/id2sentence.json', 'r') as f:
    validset_id_to_sent = json.load(f)

testset_id_to_sent['0']

len(testset_id_to_sent)

validset_id_to_sent['0']

len(validset_id_to_sent)

# Feature2id and id2feature
with open('../Dataset/ratebeer/small_500/train/feature/feature2id.json', 'r') as f:
    feature_to_id = json.load(f)
with open('../Dataset/ratebeer/small_500/train/feature/id2feature.json', 'r') as f:
    id_to_feature = json.load(f)

# user2feature
with open('../Dataset/ratebeer/small_500/train/user/user2feature.json', 'r') as f:
    user_to_feature = json.load(f)

user_feature_cnt = []
for key,value in user_to_feature.items():
    user_feature_cnt.append(len(value))
print("Total number of user on TRAIN set: {}".format(len(user_feature_cnt)))
print("Mean number of feature per user on train set: {}".format(np.mean(user_feature_cnt)))
print("Max number of feature per user on train set: {}".format(np.max(user_feature_cnt)))
print("Min number of feature per user on train set: {}".format(np.min(user_feature_cnt)))

# item2feature
with open('../Dataset/ratebeer/small_500/train/item/item2feature.json', 'r') as f:
    item_to_feature = json.load(f)

item_feature_cnt = []
for key,value in item_to_feature.items():
    item_feature_cnt.append(len(value))
print("Total number of item on TRAIN set: {}".format(len(item_feature_cnt)))
print("Mean number of feature per item on train set: {}".format(np.mean(item_feature_cnt)))
print("Max number of feature per item on train set: {}".format(np.max(item_feature_cnt)))
print("Min number of feature per item on train set: {}".format(np.min(item_feature_cnt)))

# sentence2feature
with open('../Dataset/ratebeer/small_500/train/sentence/sentence2feature.json', 'r') as f:
    sentence_to_feature = json.load(f)

sentence_feature_cnt = []
for key,value in sentence_to_feature.items():
    sentence_feature_cnt.append(len(value))
print("Total number of sentence on TRAIN set: {}".format(len(sentence_feature_cnt)))
print("Mean number of feature per sentence on train set: {}".format(np.mean(sentence_feature_cnt)))
print("Max number of feature per sentence on train set: {}".format(np.max(sentence_feature_cnt)))
print("Min number of feature per sentence on train set: {}".format(np.min(sentence_feature_cnt)))

# useritem2sentenceids - TRAIN
with open('../Dataset/ratebeer/small_500/train/useritem2sentids.json', 'r') as f:
    useritem_to_sentids = json.load(f)

# useritem2sentids - TEST
with open('../Dataset/ratebeer/small_500/test/useritem2sentids_test.json', 'r') as f:
    testset_useritem_to_sentids = json.load(f)
print("Number of user in test set: {}".format(len(testset_useritem_to_sentids)))

# useritem2sentids - VALID
with open('../Dataset/ratebeer/small_500/valid/useritem2sentids_test.json', 'r') as f:
    validset_useritem_to_sentids = json.load(f)
print("Number of user in valid set: {}".format(len(validset_useritem_to_sentids)))

empty_label = 0
candidate_sent_ids_cnt = []
for user_chunk in list(testset_useritem_to_sentids.items()):
    user_id_str = str(user_chunk[0])
    # print(user_id_str)
    user_item_chunks = list(user_chunk[1].items())
    for item_chunk in user_item_chunks:
        item_id_str = str(item_chunk[0])
        # print(item_id_str)
        # candidate_true_sent_ids = item_chunk[1]
        candidate_sent_ids = item_chunk[1][0]
        true_sent_ids = item_chunk[1][1]
        candidate_sent_ids_cnt.append(len(candidate_sent_ids))
        if len(true_sent_ids) == 0:
            # print(item_chunk)
            empty_label += 1
print("On TEST set:")
print("Number of reviews with empty true label sentences: {}".format(empty_label))
print("Max number of candidate sentence: {}".format(np.max(candidate_sent_ids_cnt)))
print("Min number of candidate sentence: {}".format(np.min(candidate_sent_ids_cnt)))
print("Mean number of candidate sentence: {}".format(np.mean(candidate_sent_ids_cnt)))

empty_label = 0
candidate_sent_ids_cnt = []
for user_chunk in list(validset_useritem_to_sentids.items()):
    user_id_str = str(user_chunk[0])
    # print(user_id_str)
    user_item_chunks = list(user_chunk[1].items())
    for item_chunk in user_item_chunks:
        item_id_str = str(item_chunk[0])
        # print(item_id_str)
        # candidate_true_sent_ids = item_chunk[1]
        candidate_sent_ids = item_chunk[1][0]
        true_sent_ids = item_chunk[1][1]
        candidate_sent_ids_cnt.append(len(candidate_sent_ids))
        if len(true_sent_ids) == 0:
            # print(item_chunk)
            empty_label += 1
print("On VALID set:")
print("Number of reviews with empty true label sentences: {}".format(empty_label))
print("Max number of candidate sentence: {}".format(np.max(candidate_sent_ids_cnt)))
print("Min number of candidate sentence: {}".format(np.min(candidate_sent_ids_cnt)))
print("Mean number of candidate sentence: {}".format(np.mean(candidate_sent_ids_cnt)))

# Combine Proxy Ids

test_review_proxy_results = []
with open('../Dataset/ratebeer/small_500/test/useritem2sentids_withproxy_multilines.json', 'w') as fw:
    for i in range(28):
        file_path = './result_small_500_test/{}.json'.format(i)
        # print(file_path)
        with open(file_path) as f:
            for line in f:
                line_data = json.loads(line)
                # print(line_data)
                user_id = str(line_data['user'])
                item_id = str(line_data['item'])
                candidate_ids = line_data["candidate_ids"]
                review_ids = line_data["review_ids"]
                proxy_ids = line_data["proxy_ids"]
                test_review_proxy_results.append([user_id, item_id, candidate_ids, review_ids, proxy_ids])
                json.dump(line_data, fw)
                fw.write('\n')
print("Number of reviews in test set: {}".format(len(test_review_proxy_results)))

valid_review_proxy_results = []
with open('../Dataset/ratebeer/small_500/valid/useritem2sentids_withproxy_multilines.json', 'w') as fw:
    for i in range(28):
        file_path = './result_small_500_valid/{}.json'.format(i)
        # print(file_path)
        with open(file_path) as f:
            for line in f:
                line_data = json.loads(line)
                # print(line_data)
                user_id = str(line_data['user'])
                item_id = str(line_data['item'])
                candidate_ids = line_data["candidate_ids"]
                review_ids = line_data["review_ids"]
                proxy_ids = line_data["proxy_ids"]
                valid_review_proxy_results.append([user_id, item_id, candidate_ids, review_ids, proxy_ids])
                json.dump(line_data, fw)
                fw.write('\n')
print("Number of reviews in valid set: {}".format(len(valid_review_proxy_results)))

df_test_result = pd.DataFrame(test_review_proxy_results, columns=['user', 'item', 'cdd', 'review', 'proxy'])
df_valid_result = pd.DataFrame(valid_review_proxy_results, columns=['user', 'item', 'cdd', 'review', 'proxy'])

df_test_result

df_valid_result

group_by_user_test_result = df_test_result.groupby('user')
group_by_user_valid_result = df_valid_result.groupby('user')

print("number of user on test set: {}".format(len(group_by_user_test_result)))
print("number of user on valid set: {}".format(len(group_by_user_valid_result)))

user_cnt = 0
user_item_testset_result = dict()
for user_df_chunk in list(group_by_user_test_result):
    user_id = int(user_df_chunk[0])
    user_id_str = str(user_df_chunk[0])
    user_df = user_df_chunk[1]
    # item-level dict
    item_level_dict = dict()
    for idx, row in user_df.iterrows():
        item_id = int(row['item'])
        item_id_str = str(row['item'])
        cdd_ids = row['cdd']
        review_ids = row['review']
        proxy_ids = row['proxy']
        assert isinstance(cdd_ids[0], str)
        assert isinstance(review_ids[0], str)
        assert isinstance(proxy_ids[0], str)
        # add this into the dict
        item_level_dict[item_id_str] = [cdd_ids, review_ids, proxy_ids]
    # add item-level dict to user-level dict
    user_item_testset_result[user_id_str] = item_level_dict
    user_cnt += 1

user_cnt = 0
user_item_validset_result = dict()
for user_df_chunk in list(group_by_user_valid_result):
    user_id = int(user_df_chunk[0])
    user_id_str = str(user_df_chunk[0])
    user_df = user_df_chunk[1]
    # item-level dict
    item_level_dict = dict()
    for idx, row in user_df.iterrows():
        item_id = int(row['item'])
        item_id_str = str(row['item'])
        cdd_ids = row['cdd']
        review_ids = row['review']
        proxy_ids = row['proxy']
        assert isinstance(cdd_ids[0], str)
        assert isinstance(review_ids[0], str)
        assert isinstance(proxy_ids[0], str)
        # add this into the dict
        item_level_dict[item_id_str] = [cdd_ids, review_ids, proxy_ids]
    # add item-level dict to user-level dict
    user_item_validset_result[user_id_str] = item_level_dict
    user_cnt += 1

# user_item_testset_result['2165']['377']
example_user_id = "2165"
example_item_id = "377"
print('On TEST set:')
print("user: {0} \t item: {1}".format(example_user_id, example_item_id))
print("Number of candidate sentence: {0} \nNumber of true review sentence: {1}\nNumber of proxy review sentence: {2}".format(
    len(user_item_testset_result[example_user_id][example_item_id][0]),
    len(user_item_testset_result[example_user_id][example_item_id][1]),
    len(user_item_testset_result[example_user_id][example_item_id][2])))
# user_item_validset_result['2165']['377']
print('On VALID set:')
print("user: {0} \t item: {1}".format(example_user_id, example_item_id))
print("Number of candidate sentence: {0} \nNumber of true review sentence: {1}\nNumber of proxy review sentence: {2}".format(
    len(user_item_validset_result[example_user_id][example_item_id][0]),
    len(user_item_validset_result[example_user_id][example_item_id][1]),
    len(user_item_validset_result[example_user_id][example_item_id][2])))

# user_item_testset_result['2165']['377']
example_user_id = "2030"
example_item_id = "548"
print('On TEST set:')
print("user: {0} \t item: {1}".format(example_user_id, example_item_id))
print("Number of candidate sentence: {0} \nNumber of true review sentence: {1}\nNumber of proxy review sentence: {2}".format(
    len(user_item_testset_result[example_user_id][example_item_id][0]),
    len(user_item_testset_result[example_user_id][example_item_id][1]),
    len(user_item_testset_result[example_user_id][example_item_id][2])))
# user_item_validset_result['2165']['377']
print('On VALID set:')
print("user: {0} \t item: {1}".format(example_user_id, example_item_id))
print("Number of candidate sentence: {0} \nNumber of true review sentence: {1}\nNumber of proxy review sentence: {2}".format(
    len(user_item_validset_result[example_user_id][example_item_id][0]),
    len(user_item_validset_result[example_user_id][example_item_id][1]),
    len(user_item_validset_result[example_user_id][example_item_id][2])))

# save this into json file
with open('../Dataset/ratebeer/small_500/test/useritem2sentids_withproxy.json', 'w') as f:
    json.dump(user_item_testset_result, f)

# save this into json file
with open('../Dataset/ratebeer/small_500/valid/useritem2sentids_withproxy.json', 'w') as f:
    json.dump(user_item_validset_result, f)
